+++ pythonk/ast/py.py
@@ -1,4 +1,4 @@
-from pythonk.ast.lib.internal_operation import InternalOperation
+from pythonk.ast.internal_operation import InternalOperation
 
 from pythonk.ast.utils.str import StrUtils
 from pythonk.parser.internal.err_handling import ArgumentTypeError
