+++ pythonk/parser/internal/basic.py
@@ -1,4 +1,4 @@
-from pythonk.ast.lib.basic import BasicOperation
+from pythonk.ast.basic import BasicOperation
 from pythonk.parser.internal.standard_grammar import StandardGrammar
